♻️ : Split src/commands into interfaces and submodule of implementations
