+++ src/command/commands/saul.rs
@@ -8,9 +8,9 @@ use coap_message::MinimalWritableMessage;
 use minicbor::Decoder;
 use minicbor::Encoder;
 
-use crate::commands::Command;
-use crate::commands::CommandHandler;
-use crate::commands::CommandRegistry;
+use super::Command;
+use super::CommandHandler;
+use super::CommandRegistry;
 
 /// This is an example on how to use cbor as payload for the coap request.
 #[derive(Parser, Debug)]
